Iteration 2

+++ backend/api/roster.controller.js
@@ -1,20 +1,80 @@
 const { date } = require("joi")
 const Joi = require("joi")
 const RostersList = require("../models/roster.js")
-const { addRosterListValidation, rosterQueryValidation, massCreateValidation, deleteRosterValidation } = require("../validation.js")
+const { addRosterListValidation, rosterQueryValidation, massCreateValidation, deleteRosterValidation, laterRosterQueryValidation } = require("../validation.js")
 const formidable = require('formidable');
 const XLSX = require("xlsx");
 const { convert_to_json, createNewRoster, findTypes, editRoster, appendRoster } = require("./roster.handlers.js");
 const { createTrail } = require("./auditTrail.controller")
-const ConfigCtrl = require("./config.controller.js");
+const { ConfigController } = require("./config.controller.js");
 const { ConnectionStates } = require("mongoose");
 
 class RosterController {
 
+    static async viewRoster(req, res, next) {
+
+        const validationError = rosterQueryValidation(req.query).error
+        if (validationError)
+            return res.status(400).json({ message: validationError.details[0].message })
+        const weekday = ["Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"];
+
+        let date = new Date()
+        if (req.query.date)
+            date = new Date(req.query.date)
+
+        const day = new Date(date.toDateString())
+        let result = await RostersList.findOne({ date: day })
+        if (!req.query.board)
+            return res.json(result.rosters)
+
+        const assignments = await ConfigController.getBoardAssignments(req.query.board)
+        if (!assignments)
+            return res.status(400).json({ message: "Config not found" })
+
+        if (!result)
+            return res.json({
+                rosters: [], timeString: new Date().toLocaleTimeString(),
+                dateString: weekday[date.getDay()] + ", " + date.toLocaleDateString('en-GB')
+            })
+
+        for (let i = 0; i < result.rosters.length; i++) {
+            const roster = result.rosters[i]
+            const assignmentSet = new Set(assignments[roster.staffType])
+            roster.roster = roster.roster.filter((staff) => {
+                const obj = JSON.parse(JSON.stringify(staff))
+                const count = Object.values(obj).length
+                return (assignmentSet.has(staff.assignment.trim()) && count > 1)
+            })
+        }
+
+
+        return res.json({
+            rosters: result.rosters, timeString: new Date().toLocaleTimeString(),
+            dateString: weekday[date.getDay()] + ", " + date.toLocaleDateString('en-GB')
+        })
+    }
+
+
+    static async viewLaterRoster(req, res, next) {
+
+        const validationError = laterRosterQueryValidation(req.query).error
+        if (validationError)
+            return res.status(400).json({ message: validationError.details[0].message })
+
+        let date = new Date()
+        if (req.query.date)
+            date = new Date(req.query.date)
+        else
+            date.setDate(date.getDate() - 1)
+        let result = await RostersList.find({ date: { $gte: date } }).sort({ date: 1 })
+        return res.json(result)
+
+    }
+
     static async ExceltoJson(req, res, next) {
         const form = new formidable.IncomingForm({ multiples: true });
         let rosters = []
-        form.parse(req, (err, fields, files) => {
+        form.parse(req, async (err, fields, files) => {
 
             const entries = files["Upload Excel"].length ? files["Upload Excel"] : [files["Upload Excel"]]
             // Loop through all workbooks
@@ -27,10 +87,15 @@ class RosterController {
                 //Loop through all sheets
                 for (let j = 0; j < workbook.SheetNames.length; j++) {
                     const name = workbook.SheetNames[j]
-                    roster = convert_to_json(workbook.Sheets[name], date)
 
-                    console.log(roster)
+                    try {
+                        roster = await convert_to_json(workbook.Sheets[name])
 
+                    } catch (e) {
+                        console.log(e)
+                        return res.status(400).json({ message: "Error converting excel", rosters: [] })
+
+                    }
                     const validationError = addRosterListValidation({ username: "admin", ...roster }).error
                     if (validationError)
                         return res.status(400).json({ message: validationError.details[0].message, rosters: [] })
@@ -46,51 +111,13 @@ class RosterController {
 
     }
 
-    static async viewRoster(req, res, next) {
-
-        const validationError = rosterQueryValidation(req.query).error
-        if (validationError)
-            return res.status(400).json({ message: validationError.details[0].message })
-        const weekday = ["Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"];
-
-        let date = new Date()
-        if (req.query.date)
-            date = new Date(req.query.date)
-
-        let result = await RostersList.findOne({ date: date })
-        if (!req.query.board)
-            return res.json(result.rosters)
-
-        const assignments = await ConfigCtrl.getBoardAssignments(req.query.board)
-        if (!assignments)
-            return res.status(400).json({ message: "Config not found" })
-
-        if (!result)
-            return res.json({
-                rosters: [], timeString: new Date().toLocaleTimeString(),
-                dateString: weekday[date.getDay()] + ", " + date.toLocaleDateString('en-GB')
-            })
-
-        for (let i = 0; i < result.rosters.length; i++) {
-            const roster = result.rosters[i]
-
-            const assignmentSet = new Set(assignments[roster.staffType])
-            roster.roster = roster.roster.filter((staff) => { return assignmentSet.has(staff.assignment.trim()) })
-        }
-
-
-        return res.json({
-            rosters: result.rosters, timeString: new Date().toLocaleTimeString(),
-            dateString: weekday[date.getDay()] + ", " + date.toLocaleDateString('en-GB')
-        })
-    }
-
     static async massCreateRoster(req, res, next) {
+
         const validationError = massCreateValidation(req.body).error
         if (validationError)
             return res.status(400).json({ message: validationError.details[0].message })
 
-        const rosters = req.body.rosters
+        const rosters = req.body.rostersList
         const username = req.body.username
         for (let i = 0; i < rosters.length; i++) {
             const roster = rosters[i]
@@ -118,9 +145,15 @@ class RosterController {
     }
 
     static async createRoster(req, res, next) {
+
         const validationError = addRosterListValidation(req.body).error
         if (validationError)
             return res.status(400).json({ message: validationError.details[0].message })
+        const takenDate = await RostersList.findOne({ date: req.body.date })
+
+        if (takenDate) {
+            return res.status(400).json({ message: "Date is already created" })
+        }
 
         try {
             await createNewRoster(req.body.username, req.body.date, req.body.rosters)
@@ -128,6 +161,7 @@ class RosterController {
         catch (err) {
             return res.status(500).json({ message: err.message })
         }
+
 
         return res.json({ message: "Roster Created" })
 
@@ -144,34 +178,14 @@ class RosterController {
         return res.json(types)
     }
 
-    static async searchName(req, res, next) {
-
-        const query = { date: req.query.date, staffType: req.query.name }
-        const validationError = rosterQueryValidation(query).error
-        if (validationError)
-            return res.status(400).json({ message: validationError.details[0].message })
-        const result = await RostersList.aggregate([
-            //Some date check 
-            { $unwind: "$rosters" },
-            { $unwind: "$rosters.roster" },
-            {
-
-                $match: {
-                    $or: [{ "rosters.roster.am.name": "ZX NG" }, { "rosters.roster.pm.name": "ZX NG" }]
-                }
-            },
-            { $replaceWith: "$rosters" },
-        ])
-
-        return res.json({ message: "Searched", result: result })
-    }
-
     static async deleteRoster(req, res, next) {
-
         const validationError = deleteRosterValidation(req.body).error
         if (validationError)
             return res.status(400).json({ message: validationError.details[0].message })
-        const result = await RostersList.deleteOne({ date: new Date(req.body.date) }).exec();
+        if (req.user.username !== req.body.username)
+            return res.status(400).json({ message: "User is not authorised to perform this action" })
+
+        await RostersList.deleteOne({ date: new Date(req.body.date) }).exec();
         try {
             createTrail({
                 username: req.body.username, type: "delete-roster", deletedDocumentDate: req.body.date
@@ -184,4 +198,6 @@ class RosterController {
     }
 
 }
+
+
 module.exports = RosterController
